requirements.txt
numpy
pandas
seaborn
matplotlib
psycopg2
python-dotenv
plotly
nltk
streamlit